Hello, this is the bot-version 1.2.0. This version contains a translator from the datastructure of this agent to the datastructure of the AKI-RealBot from the TU-Darmstadt. Whereupon the datastructure of the AKI-RealBot was reduced to the necessary part for using the strategy of it. Furthermore has this bot-version a new strategy. The new strategy is the strategy of the AKI-RealBot. With the exception if the bot does not know all players (the case is sometimes in the pre-flop-phase), it is used the strategy of the previous version strategy.strategyPokerStrategy.StrategyOne. In this version is the stack of every player missed for completely using the strategy of the AKI-RealBot. The stack of every player has to be determined by an OCR library but there are not many OCR libraries for free! :( au revoir, joschka

+++ PokerBot/src/strategy/strategyPokerChallenge/simulation/real/datastructure/RealPlayerState.java
@@ -13,13 +13,13 @@ public class RealPlayerState {
 	private Card[] hole;
 	private boolean active;
 	private boolean canAct;
-	private int amountIn;
-	private int amountBet;
+	private double amountIn;
+	private double amountBet;
 	private int seat;
 	private int playerSeat;
 	private int bucket;
 	
-	public RealPlayerState(boolean active, boolean canRaiseNextRound, int inPot, int seat, int playerSeat) {
+	public RealPlayerState(boolean active, boolean canRaiseNextRound, double inPot, int seat, int playerSeat) {
 		this.hole = null;
 		this.active = active;
 		this.bucket = -1;
@@ -70,14 +70,14 @@ public class RealPlayerState {
 	/**
 	 * @return the amountIn
 	 */
-	public int getAmountIn() {
+	public double getAmountIn() {
 		return amountIn;
 	}
 
 	/**
 	 * @param amountIn the amountIn to set
 	 */
-	public void setAmountIn(int amountIn) {
+	public void setAmountIn(double amountIn) {
 		this.amountIn = amountIn;
 	}
 
@@ -85,7 +85,7 @@ public class RealPlayerState {
 	/**
 	 * @return the amountBet
 	 */
-	public int getAmountBet() {
+	public double getAmountBet() {
 		return amountBet;
 	}
 
@@ -110,7 +110,7 @@ public class RealPlayerState {
 		this.canAct = canRaiseNextRound;
 	}
 
-	public boolean canAct(int maxAmount) {
+	public boolean canAct(double maxAmount) {
 		return isActive() && (canAct || amountIn<maxAmount);
 	}
